+++ documentation/src/components/ExampleTextarea.vue
@@ -2,6 +2,7 @@
 import { Textarea as VanillaTextArea } from '@flavorly/vanilla-components'
 import { ref } from 'vue'
 const value = ref(null)
+const valueDisabled = ref(null)
 const valueErrors = ref(null)
 </script>
 
@@ -20,6 +21,13 @@ const valueErrors = ref(null)
         placeholder="You can also have a long story about Elon Musk memes on Twitter"
         feedback="Im showing only because you cleared errors"
         errors="We have a problem Musk, Abort! Write to clear"
+      />
+      <!-- Disabled -->
+      <VanillaTextArea
+        v-model="valueDisabled"
+        :disabled="true"
+        placeholder="Hey! Im disabled and you should be able to touch me :p"
+        feedback="Im disabled and shouldnt be changed"
       />
     </div>
   </PreviewWrapper>